추가: article 유효성 검사
